refactor: extract cheat system to minarch_cheats.c/h
Move the cheat loading, parsing, and path-resolution code (lines 388–745)
from the minarch.c monolith into dedicated minarch_cheats.c/h files.
Promote getAlias() and Cheat_getPaths() from static to file-visible
symbols declared in the shared internal headers so both translation units
can access them.
